examples/basic/requirements.txt
pyzmq
msgpack
PyJWT